refactor: 인피니티 스크롤 위치 관련 코드 고치는 중
