+++ src/main/java/com/example/portfolio/controller/ProjectController.java
@@ -28,7 +28,6 @@ import com.example.portfolio.dto.CategoryUpdateDto;
 import com.example.portfolio.dto.PhotoListDto;
 import com.example.portfolio.dto.ProjectCreateDto;
 import com.example.portfolio.dto.ProjectDetailDto;
-import com.example.portfolio.dto.ProjectDetailPageDto;
 import com.example.portfolio.dto.ProjectListCustomDto;
 import com.example.portfolio.dto.ProjectListDto;
 import com.example.portfolio.dto.ProjectUpdateDto;
@@ -171,7 +170,7 @@ public class ProjectController {
 	}
 
 	@GetMapping("/photos/{id}")
-	public ProjectDetailPageDto getPhotos(
+	public Slice<PhotoListDto> getPhotos(
 			@PageableDefault( size = 10) Pageable pageable,
 			@PathVariable("id") Long projectId) {
 		return projectService.getPhotoList(pageable, projectId);
